fix(options): use watchCollection instead of simple watch to watch options collection
Using a simple watch to watch the options collection will cause a 'digest overflow' if a filter is used in options expression. The filter returns a new array on each digest lifecycle and the watch will fire each time. Using a watchCollection solves this issue.
watchCollection also has the advantage to be informed if an element is added to the collection. So the user of the directive won't have to assign a new array to update the options anymore.

Closes #6