fix(options): mark options as selected on change of options even if options are objects
Internal Option representation is an object which is created newly on each change of options collection. Array#indexOf with objects compares by reference and not equality. Because of that the option put into ngModel out site are not marked as selected.
By calculating and comparing hash codes of the option the directive is now independent of the kind of options.

Closes #7